+++ src/pyDE1/dispatcher/dispatcher.py
@@ -26,7 +26,6 @@ from pyDE1.dispatcher.implementation import get_resource_to_dict, \
 from pyDE1.scale.processor import ScaleProcessor
 from pyDE1.supervise import SupervisedTask
 
-READ_BACK_ON_PATCH = False
 
 QUEUE_TOO_DEEP = 1  # If deeper than this, something is probably wrong, log
 
@@ -119,7 +118,7 @@ async def _request_queue_processor(request_queue: asyncio.Queue,
             )
 
         elif got.method is HTTPMethod.PATCH:
-            resource_dict = None
+            results_list = None
             try:
                 # TODO: Should be "ready" and not just "connected"
                 if (got.connectivity_required['DE1'] and not de1.is_connected):
@@ -128,25 +127,24 @@ async def _request_queue_processor(request_queue: asyncio.Queue,
                         and scale_processor.scale is not None
                         and not scale_processor.scale.is_connected):
                     raise DE1NotConnectedError("Scale not connected")
-                await patch_resource_from_dict(got.resource,
-                                               got.payload)
+                results_list = await patch_resource_from_dict(got.resource,
+                                                              got.payload)
             except Exception as e:
                 exception = e
                 logger.error(
                     f"Exception in processing {got.method} {got.resource}"
                     f" {repr(exception)}")
-            if READ_BACK_ON_PATCH and \
-                    exception is None and got.resource.can_get:
-                resource_dict = await get_resource_to_dict(got.resource)
+
             response = APIResponse(
                 original_timestamp=got.timestamp,
                 timestamp=time.time(),
-                payload = resource_dict,
+                payload = results_list,
                 exception=exception,
             )
 
         elif got.method is HTTPMethod.PUT \
                 and got.resource is Resource.DE1_PROFILE:
+            results_list = None
             try:
                 # TODO: Should be "ready" and not just "connected"
                 if (got.connectivity_required['DE1'] and not de1.is_connected):
@@ -157,18 +155,19 @@ async def _request_queue_processor(request_queue: asyncio.Queue,
                     raise DE1NotConnectedError("Scale not connected")
 
                 # TODO: Implement PUT properly (check for completeness)
-                await patch_resource_from_dict(got.resource,
-                                               got.payload)
+                results_list = await patch_resource_from_dict(got.resource,
+                                                              got.payload)
             except Exception as e:
                 exception = e
                 logger.error(
                     f"Exception in processing {got.method} {got.resource}"
                     f" {repr(exception)}")
 
+
             response = APIResponse(
                 original_timestamp=got.timestamp,
                 timestamp=time.time(),
-                payload = None,
+                payload = results_list,
                 exception=exception,
             )